feat: 클린 아키텍처 적용
