+++ src/main/java/lime1st/netty/auth/application/port/in/TokenExpireUseCase.java
@@ -1,0 +1,8 @@
+package lime1st.netty.auth.application.port.in;
+
+import reactor.core.publisher.Mono;
+
+public interface TokenExpireUseCase {
+
+    Mono<Void> tokenExpire(String token);
+}
