+++ src/main/java/lime1st/netty/auth/application/port/in/TokenVerifyUseCase.java
@@ -1,0 +1,8 @@
+package lime1st.netty.auth.application.port.in;
+
+import reactor.core.publisher.Mono;
+
+public interface TokenVerifyUseCase {
+
+    Mono<String> verifyToken(String token);
+}
