+++ src/main/java/lime1st/netty/auth/domain/Token.java
@@ -1,0 +1,13 @@
+package lime1st.netty.auth.domain;
+
+public record Token(
+        long userId,
+        String email,
+        long issuedDate,
+        long expireDate
+) {
+
+    public boolean isExpired() {
+        return System.currentTimeMillis() > expireDate;
+    }
+}
