--- src/main/java/lime1st/netty/bootstrap/ApiServerInitializer.java
@@ -1,30 +1,0 @@
-package lime1st.netty.bootstrap;
-
-import io.netty.channel.ChannelInitializer;
-import io.netty.channel.ChannelPipeline;
-import io.netty.channel.socket.SocketChannel;
-import io.netty.handler.codec.http.HttpContentCompressor;
-import io.netty.handler.codec.http.HttpObjectAggregator;
-import io.netty.handler.codec.http.HttpRequestDecoder;
-import io.netty.handler.codec.http.HttpResponseEncoder;
-import lime1st.netty.api.handler.ApiRequestParser;
-import lime1st.netty.service.common.Router;
-
-public class ApiServerInitializer extends ChannelInitializer<SocketChannel> {
-
-    private final Router router;
-
-    public ApiServerInitializer(Router router) {
-        this.router = router;
-    }
-
-    @Override
-    protected void initChannel(SocketChannel ch) throws Exception {
-        ChannelPipeline p = ch.pipeline();
-        p.addLast(new HttpRequestDecoder());
-        p.addLast(new HttpObjectAggregator(65536));
-        p.addLast(new HttpResponseEncoder());
-        p.addLast(new HttpContentCompressor());
-        p.addLast(new ApiRequestParser(router));
-    }
-}
