+++ src/main/java/lime1st/netty/server/application/port/in/ApiRequest.java
@@ -1,0 +1,9 @@
+package lime1st.netty.server.application.port.in;
+
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import reactor.core.publisher.Mono;
+
+public interface ApiRequest {
+
+    Mono<ObjectNode> executeService();
+}
