+++ src/main/java/lime1st/netty/server/framework/ApiServerInitializer.java
@@ -1,0 +1,30 @@
+package lime1st.netty.server.framework;
+
+import io.netty.channel.ChannelInitializer;
+import io.netty.channel.ChannelPipeline;
+import io.netty.channel.socket.SocketChannel;
+import io.netty.handler.codec.http.HttpContentCompressor;
+import io.netty.handler.codec.http.HttpObjectAggregator;
+import io.netty.handler.codec.http.HttpRequestDecoder;
+import io.netty.handler.codec.http.HttpResponseEncoder;
+import lime1st.netty.server.adapter.in.web.ApiRequestParser;
+import lime1st.netty.server.adapter.in.web.Router;
+
+public class ApiServerInitializer extends ChannelInitializer<SocketChannel> {
+
+    private final Router router;
+
+    public ApiServerInitializer(Router router) {
+        this.router = router;
+    }
+
+    @Override
+    protected void initChannel(SocketChannel ch) throws Exception {
+        ChannelPipeline p = ch.pipeline();
+        p.addLast(new HttpRequestDecoder());
+        p.addLast(new HttpObjectAggregator(65536));
+        p.addLast(new HttpResponseEncoder());
+        p.addLast(new HttpContentCompressor());
+        p.addLast(new ApiRequestParser(router));
+    }
+}
